% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_function.R
\name{add_rowdata_detection}
\alias{add_rowdata_detection}
\title{add_rowdata_detection}
\usage{
add_rowdata_detection(pe_, design, assay)
}
\arguments{
\item{pe_}{q_feature object where to add rowdata}

\item{design}{EDF data frame}

\item{assay}{assay name where to add the statistics}
}
\value{
pe_: modified q-feature object
}
\description{
This function adds number and percentage of non missing value for each group in assay specified.
Moreover it adds computed the number of peptide per protien and the total number of non missing values.
}
\author{
Andrea Argentini
}
