% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_function.R
\name{check_columns_presence}
\alias{check_columns_presence}
\title{check_columns_presence}
\usage{
check_columns_presence(df, min_features)

check_columns_presence(df, min_features)
}
\arguments{
\item{df}{data frame containing experiment design data}

\item{min_features}{vector with the column names to be checked}
}
\value{
status  int 0 / 1 error found

type_raw format file of the raw file detected ,

error error message

status : int 0 / 1 error found

type_raw: format file of the raw file detected ,

error: error message
}
\description{
This function checks the present of a set of features in the columns name
of the input data frame.

This function checks some main sanity controls in the design file
e.g. Name of columns, min available columns.
}
\author{
Andrea Argentini
}
