% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_function.R
\name{check_length_design_data}
\alias{check_length_design_data}
\title{check_length_design_data}
\usage{
check_length_design_data(data_, design)
}
\arguments{
\item{data_}{data frame containing the DIA-NN report data}

\item{design}{data frame containing experiment design data}
}
\value{
status  int 0 / 1: error found, 2: samples in DIA-NN report are more than samples in experiment design data

error error message

message: message returned if data frame containing the DIA-NN report data is filtered
}
\description{
This function checks the names and the number of samples in DIA-NN report and experiment design data,
if DIA-NN report has more samples than design file, only sample present in design file are kept.
Remark : Model result are supposed to be in proteinRS layer.
}
\author{
Andrea Argentini
}
