% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_function.R
\name{checkVariables}
\alias{checkVariables}
\title{checkVariables}
\usage{
checkVariables(inputParams, dfDesign, variables)
}
\arguments{
\item{inputParams}{comparisons groups in input parameter list}

\item{dfDesign}{data frame containing experiment design data}

\item{variables}{data frame containing experiment design data}
}
\value{
status  int 0 / 1 error message

error
}
\description{
This function implement sanity check if the values are correct with respect
to the design file
}
\author{
Andrea Argentini
}
