% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_function.R
\name{dep_volcano}
\alias{dep_volcano}
\title{dep_volcano}
\usage{
dep_volcano(label, data, params, layer, annotation_ev)
}
\arguments{
\item{label}{Character. Contrast label identifying the result column}

\item{data}{QFeatures object with proteomics data.}

\item{params}{List. Contains analysis parameters like thresholds and ensembl settings.}

\item{layer}{Character. The name of the layer (e.g., "proteinRS", "peptideNorm").#' @return toptable result as dataframe}

\item{annotation_ev}{Logical. Whether to include annotation columns.}
}
\value{
A list with the following elements:
\describe{
  \item{toptable}{Data.frame of differential expression results.}
  \item{volcano}{ggplot object of the volcano plot.}
  \item{volcano2file}{ggplot object with annotated labels for saving/export.}
  }
}
\description{
Generates volcano plots and returns both raw and annotated differential expression results for a given contrast. Assumes model results are stored in the proteinRS or peptideNorm layer.
}
\author{
Andrea Argentini
}
