% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_function.R
\name{dfTowideDIANN_20}
\alias{dfTowideDIANN_20}
\title{dfTowideDIANN_20}
\usage{
dfTowideDIANN_20(data, precursorquan, mbr, wide_colums)
}
\arguments{
\item{data}{frame containing the DIA-NN report data}

\item{precursorquan}{columns to use to pivot into a wide format}

\item{mbr}{mbr parameter}

\item{wide_colums}{List of the columns that should be attached to the result}
}
\value{
status : A data frame of DIA-NN result in a wide format
}
\description{
This function after some quality check filters DIA-NN result using Q-value and
and PG Qvalues.
It makes a wide version of the DIA-NN result using the precursorquant information
}
\author{
Andrea
}
