% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_function.R
\name{filteringNA_qfeat_ev}
\alias{filteringNA_qfeat_ev}
\title{filteringNA_qfeat_ev}
\usage{
filteringNA_qfeat_ev(pe_, params, design)
}
\arguments{
\item{pe_}{q_feature object}

\item{params}{EDF data frame}

\item{design}{EDF data frame}
}
\value{
status: int 0 non error  / 1 error found

q_feat: q-feature created/ modified

error: error message
}
\description{
This function adds number and percentage of non missing value for each group in assay specified.
Moreover it adds computed the number of peptide per protein and the total number of non missing values.
}
\author{
Andrea Argentini
}
