% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_function.R
\name{generate_pca_plots}
\alias{generate_pca_plots}
\title{Generate PCA Plots}
\usage{
generate_pca_plots(pe, params, layer, test_skip = FALSE)
}
\arguments{
\item{pe}{A QFeatures object containing the data for PCA.}

\item{params}{A list containing parameters for the run, including `PCA_comparison`
(a character vector of variables to use for PCA) and `folder_prj` (the output folder).}

\item{layer}{The layer of the QFeatures object to use for PCA.}

\item{test_skip}{Logical. If `TRUE`, skips saving plots as PDF files.}
}
\value{
A list containing:
  - `plots`: A list of ggplot objects for each PCA plot.
  - `msgs`: A list of messages indicating the type of PCA performed.
  - `status`: A list of statuses (0 for success, 1 for invalid combinations).
}
\description{
This function generates PCA plots from a list of confounders or variables
in the `colData` of a QFeatures object. It supports combinations of
one or two variables for color and shape, with specific type combinations:
- Character | Factor and Character | Factor
- Numerical and Character | Factor
- Character | Factor (single variable)
}
\details{
The function returns a list of ggplot objects and saves the plots as PDFs.
}
