% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_function.R
\name{get_DEall_columns}
\alias{get_DEall_columns}
\title{get_DEall_columns}
\usage{
get_DEall_columns(layer, annotation_ev, perc_field)
}
\arguments{
\item{layer}{Character. Either "proteinRS" or "peptideNorm".}

\item{annotation_ev}{Logical. Whether to include annotation columns like Subcellular.Location.}

\item{perc_field}{Character vector of percentage fields to include.}
}
\value{
A character vector of column names to extract from the results
}
\description{
Returns a set of column names for the final output table, based on layer type and annotation presence.
}
\author{
Andrea Argentini
}
