% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_function.R
\name{get_peptide_data}
\alias{get_peptide_data}
\title{get_peptide_data}
\usage{
get_peptide_data(data, label, perc_field, annotation_ev)
}
\arguments{
\item{data}{QFeatures object containing proteomics data.}

\item{label}{Character. Column name in the peptide results layer to extract.}

\item{perc_field}{Character vector. Column names matching percentage values to include.}

\item{annotation_ev}{Logical. Whether to include Subcellular.Location annotation.}
}
\value{
A data.frame with extracted and joined peptide-level results.#' @importFrom dplyr select left_join
}
\description{
Helper function to extract and join relevant peptide-level data, including annotations and percentage fields, from a QFeatures object.
}
\author{
Andrea Argentini
}
