% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_function.R
\name{import2_qfeature}
\alias{import2_qfeature}
\title{import2_qfeature}
\usage{
import2_qfeature(
  diaNN_data,
  design,
  params,
  min_col_need_design,
  diann_colname,
  task
)
}
\arguments{
\item{diaNN_data}{data frame containing the DIA-NN in a wide format}

\item{design}{design information as data frame}

\item{params}{list of parameters}

\item{min_col_need_design}{list of the mandatory fields in EDF file}

\item{diann_colname}{columns names}

\item{task}{string that indicated if it is for export  qfeat or de analysis}
}
\value{
status int 0 non error  / 1 error found

q_feat q-feature created/ modified

error error message
}
\description{
This function create an q-feature obj, containing the DIA-NN precursor intensities in the assay precursor.
ColData is also set with the experiment design information. Several sanity checks are performed like:
1)  columns present in design file
2)  check number and  sample name found in the DIA-NN report and in the design file
3) check confounder variable present in the design file
Finally name matching between run in DIA-report andin the design file is done,
using params$wildstr_run to select the run name among the the other information columns
}
\author{
Andrea Argentini
}
