% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_function.R
\name{make_annotated_volcano}
\alias{make_annotated_volcano}
\title{make_annotated_volcano}
\usage{
make_annotated_volcano(df, params, title, label_col = "label_DE")
}
\arguments{
\item{df}{Data.frame with differential expression results and a label column.}

\item{params}{List of parameters including thresholds for fold change and adjusted p-value.}

\item{title}{Character. Title for the plot}

\item{label_col}{Character. Column name used for labeling (default: "label_DE").}
}
\value{
A ggplot object with text labels added.
}
\description{
Generates a volcano plot with gene or peptide labels for significantly expressed features.
}
\author{
Andrea Argentini
}
