% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_function.R
\name{msqrob_model}
\alias{msqrob_model}
\title{msqrob_model}
\usage{
msqrob_model(pe_, params, layer)
}
\arguments{
\item{pe_}{q_feature object where to perfomr DE analsysis}

\item{params}{parameters}

\item{layer}{name of the layer of the Qfeature obj}
}
\value{
error: modified q-feature object
}
\description{
This function perform DE Analysis with the input formula and the comparison of interest
}
\author{
Andrea Argentini
}
