% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_function.R
\name{process_results}
\alias{process_results}
\title{process_results}
\usage{
process_results(all_res, params)
}
\arguments{
\item{all_res}{Data.frame containing statistical test results.}

\item{params}{List. Must include `FC_thr`, `adjpval_thr`, and `filt_NaNDE`.}
}
\value{
A data.frame with an additional "differential_expressed" column.
}
\description{
Adds a "differential_expressed" column to indicate UP, DOWN, or NO differential expression status based on logFC and adjusted p-value thresholds.
}
\author{
Andrea Argentini
}
