% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_function.R
\name{processing_qfeat_protein}
\alias{processing_qfeat_protein}
\title{processing_qfeat_protein}
\usage{
processing_qfeat_protein(pe_, params, aggr_mth_fun)
}
\arguments{
\item{pe_}{q_feature object where to add rowdata}

\item{params}{EDF data frame}

\item{aggr_mth_fun}{function for protein summarization}
}
\value{
status int 0 non error  / 1 error found

q_feat q-feature created / modified

error error message
}
\description{
This function applies the following steps from precursor assays:
1. Log2 transformation of the percursor intensities (assay name: precursorLog)
2. Normalization of log2transformed intensities based on params$normalization method (assay name: precursorNorm).
3. Protein summarization based on  aggr_method_f method (assay name: proteinRS)
}
\author{
Andrea Argentini
}
