% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_function.R
\name{read_DIANN_report}
\alias{read_DIANN_report}
\title{read_DIANN_report}
\usage{
read_DIANN_report(params)
}
\arguments{
\item{params}{q_feature}
}
\value{
status int 0 non error  / 1 error found

error error message is any

w_diann diann data in a wide format

df_design design data as data.frame
}
\description{
bla bla
}
\author{
Andrea Argentini
}
