% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_template.R
\name{render_dia_report}
\alias{render_dia_report}
\title{Render a DIA-NN report using a Quarto template}
\usage{
render_dia_report(params_report, template, report_folder, report_filename)
}
\arguments{
\item{params_report}{parameters}

\item{template}{template name}

\item{report_folder}{description}

\item{report_filename}{output report file}
}
\value{
The full path to the rendered report.
}
\description{
Render a DIA-NN report using a Quarto template
}
\details{
The `params` list must contain the following elements:
\describe{
  \item{\code{report_target_folder}}{The target folder where the report and parameters will be saved.}
  \item{\code{input_file}}{The path to the input file.}
  \item{\code{design_file}}{The path to the design file.}
  \item{\code{title}}{The title of the report.}
  \item{\code{subtitle}}{The subtitle of the report.}
  \item{\code{author}}{The author of the report.}
  \item{\code{description}}{A description of the report.}
  \item{\code{contrast}}{The contrast for the analysis.}
  \item{\code{aggr_method}}{The aggregation method.}
  \item{\code{normalization}}{The normalization method.}
  \item{\code{formula}}{The formula for the analysis.}
  \item{\code{Proteotypic}}{Whether to use proteotypic peptides.}
  \item{\code{pep_per_prot}}{The minimum number of peptides per protein.}
  \item{\code{nNonZero}}{The minimum number of non-zero values.}
  \item{\code{confounder_list}}{A vector of confounders.}
  \item{\code{PCA_comparison}}{A vector of comparisons for PCA plots.}
  \item{\code{comparisons}}{A vector of comparisons for differential analysis.}
  \item{\code{FC_thr}}{The fold change threshold.}
  \item{\code{filtering_contaminant}}{Whether to filter contaminants.}
  \item{\code{quantitative_features}}{The quantitative feature to use from DIA-NN output.}
  \item{\code{filtPerGroup}}{Whether to filter per group.}
  \item{\code{mbr}}{Whether MBR was used in DIA-NN.}
  \item{\code{template_file}}{The name of the Quarto template file.}
  \item{\code{output_filename}}{The name of the output HTML file.}
  \item{\code{DIANN_ver2}}{Whether to read DIANN version > 2 output.}
}
}
\author{
andrea Argentini
}
