% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_function.R
\name{select_samples_comparison}
\alias{select_samples_comparison}
\title{select_samples_comparison}
\usage{
select_samples_comparison(test_parsing, pe, variable_names)
}
\arguments{
\item{test_parsing}{input comparison label, text is already parsed}

\item{pe}{Q-feature object related to the analysis.}

\item{variable_names}{list of the variable name in the formula}
}
\value{
list with left and right parsed value of the comparison
}
\description{
This function takes the text parsed from the comparison label and select the right.
sample in each contrast
}
\author{
Andrea Argentini
}
