% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_pipeline.R
\name{step_add_ensembl}
\alias{step_add_ensembl}
\title{step_add_ensembl}
\usage{
step_add_ensembl(input)
}
\arguments{
\item{input}{A list containing pe and params_report.}
}
\value{
The updated input list with Ensembl annotation (if applied).
}
\description{
Step function that adds Ensembl annotation to the qfeature object if annotation is specified in params_report.
}
\author{
Andrea Argentini
}
