% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_pipeline.R
\name{step_add_rowdata_precursor}
\alias{step_add_rowdata_precursor}
\title{step_add_rowdata_precursor}
\usage{
step_add_rowdata_precursor(input)
}
\arguments{
\item{input}{A list containing pe and design.}
}
\value{
The updated input list with pe modified at precursor level.
}
\description{
Step function that adds row data detection at the precursor level using add_rowdata_detection.
}
\author{
Andrea Argentini
}
