% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_pipeline.R
\name{step_filter_na}
\alias{step_filter_na}
\title{step_filter_na}
\usage{
step_filter_na(input)
}
\arguments{
\item{input}{A list containing pe, params_report, and design.}
}
\value{
The updated input list with filtered pe.
}
\description{
Step function that applies filteringNA_qfeat to filter missing values in the qfeature object.
}
\author{
Andrea Argentini
}
