% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_pipeline.R
\name{step_import_qfeat_exportonly}
\alias{step_import_qfeat_exportonly}
\title{step_import_qfeat_exportonly}
\usage{
step_import_qfeat_exportonly(input)
}
\arguments{
\item{input}{A list containing dfMsqrob, design, and params_report.}
}
\value{
The updated input list with the pe (q_feat) object.
}
\description{
Step function that wraps the import2_qfeature function. It processes input data and
adds the q_feat object to the input list. This is not sanatize the comparisons input
}
\author{
Andrea Argentini
}
