% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_pipeline.R
\name{step_msqrob_de}
\alias{step_msqrob_de}
\title{step_msqrob_de}
\usage{
step_msqrob_de(input)
}
\arguments{
\item{input}{A list containing pe and params_report.}
}
\value{
The updated input list with de_comparison and updated pe.
}
\description{
Step function that runs msqrob_model and adds differential expression results to the input list.
}
\author{
Andrea Argentini
}
