% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_pipeline.R
\name{step_partial_result}
\alias{step_partial_result}
\title{step_partial_result}
\usage{
step_partial_result(input)
}
\arguments{
\item{input}{A list containing pe and params_report.}
}
\value{
The updated input list with de_comparison and updated pe.
}
\description{
Step function for pariati data
}
\author{
Andrea Argentini
}
