% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_pipeline.R
\name{step_processing_peptide}
\alias{step_processing_peptide}
\title{step_processing_protein}
\usage{
step_processing_peptide(input)
}
\arguments{
\item{input}{A list containing pe, params_report, and aggr_method_f.}
}
\value{
The updated input list with pe aggregated at protein level.
}
\description{
Step function that aggregates precursor to peptide sequence level
}
\author{
Andrea Argentini
}
