% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_pipeline.R
\name{step_processing_protein_ev}
\alias{step_processing_protein_ev}
\title{step_processing_protein_ev}
\usage{
step_processing_protein_ev(input)
}
\arguments{
\item{input}{A list containing pe, params_report, and aggr_method_f.}
}
\value{
The updated input list with pe aggregated at protein level.
}
\description{
Step function that aggregates qfeature object to protein level using the specified aggregation method.
}
\author{
Andrea Argentini
}
