% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_pipeline.R
\name{step_read_diann}
\alias{step_read_diann}
\title{step_read_diann}
\usage{
step_read_diann(input)
}
\arguments{
\item{input}{A list containing at least params_report.}
}
\value{
The updated input list with design and dfMsqrob.
}
\description{
Step function that reads the DIANN report and populates the workflow input list
with the design and dfMsqrob objects.
}
\author{
Andrea Argentini
}
