% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_template.R
\name{validate_params}
\alias{validate_params}
\title{Validate parameters for the DIA-NN report}
\usage{
validate_params(params)
}
\arguments{
\item{params}{A list of parameters to be validated.}
}
\value{
TRUE if all parameters are valid, otherwise stops with an error message.
}
\description{
Validate parameters for the DIA-NN report
}
