% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_template.R
\name{validate_params_qfeatobj}
\alias{validate_params_qfeatobj}
\title{Validate parameters for qfeature obj function only}
\usage{
validate_params_qfeatobj(params)
}
\arguments{
\item{params}{A list of parameters to be validated.}
}
\value{
TRUE if all parameters are valid, otherwise stops with an error message.
}
\description{
Validate parameters for qfeature obj function only
}
