% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_function.R
\name{dep_volcano}
\alias{dep_volcano}
\title{DEP_volcano}
\usage{
dep_volcano(label, data, params, layer)
}
\arguments{
\item{label}{contrast name}

\item{data}{Qfeatures object}

\item{params}{document parameters}

\item{layer}{QFeature object}
}
\value{
toptable result as dataframe

volcano volcano ggplot object

volcano2file volcano ggplot annotated
}
\description{
This function computes volcano plot and return the toptable for each comparison
Remark : Model result are supposed to be in proteinRS layer.
}
\author{
Andrea Argentini
}
