% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_template.R
\name{merge_default_parameters}
\alias{merge_default_parameters}
\title{merge_default_parameters}
\usage{
merge_default_parameters(params_int)
}
\arguments{
\item{params_int}{parameters}
}
\value{
merged parameters
}
\description{
merge_default_parameters
}
