% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_function.R
\name{partially_present}
\alias{partially_present}
\title{partially_present}
\usage{
partially_present(pe_, params, layer)
}
\arguments{
\item{pe_}{q_feature object where to perfomr DE analsysis}

\item{params}{parameters}

\item{layer}{name of the layer of the Qfeature obj}
}
\value{
error: modified q-feature object
}
\description{
This function perform DE Analysis with the input formula and the comparison of interest
This work only with one VARIABLE Design --> ~ Group
}
\author{
Andrea Argentini
}
