% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_function.R
\name{processing_qfeat_peptide}
\alias{processing_qfeat_peptide}
\title{processing_qfeat_peptide}
\usage{
processing_qfeat_peptide(pe_, params, aggr_mth_fun)
}
\arguments{
\item{pe_}{q_feature object where to add rowdata}

\item{params}{EDF data frame}

\item{aggr_mth_fun}{function for protein summarization}
}
\value{
status: int 0 non error  / 1 error found

q_feat: q-feature created / modified

error: error message
}
\description{
This function applies the following steps from precursor assay:
1. Summarize precursor to peptide (stripped sequence) intensities using sum function (assay name PeptideRawSum)
1. Log2 transformation of the peptide intensities (assay name: peptideLog)
2. Normalization of log2transformed peptide intensities based on params$normalization method (assay name: peptideNorm).
}
\author{
Andrea Argentini
}
